Creates connection_activation module
- Abstracts the CapabilitiesExchange and ConnectionFinalization
sequences into a separate connection_activation module. This allows us
to reuse them when we receive a Server Deactivate All.
- Creates IoChannelPdu enum to be used to account for the potential of
receiving a ServerDeactivateAll on the I/O channel
