+++ crates/ironrdp-web/src/session.rs
@@ -603,6 +603,7 @@ impl Session {
                             hotspot_y,
                         })?;
                     }
+                    ActiveStageOutput::DeactivateAll(_) => todo!("DeactivateAll"),
                     ActiveStageOutput::Terminate(reason) => break 'outer reason,
                 }
             }
